feat(streamd)!: Add flag listen-cidr
This adds a flag (listen-cidr), which can be used to specify the CIDR containing the IP to listen on.
This allows deploying VMPs with a unified config to listen to ay IP within assigned to the machine within our tailnet.
The configuration would look as follows:

```sh
        ${pkgs.streamd}/bin/streamd \
          -hw-accel \
	  -source-present="decklinkvideosrc" \
	  -source-present-opts="connection=hdmi device-number=0" \
	  [...]
	  -listen-cidr="10.64.0.0/10"
```

If a machines ip address were to change within the network, the configuration would still be valid.

BREAKING CHANGE:
The flags

- listen-comb-srt
- listen-present-srt
- listen-camera-srt

have been removed in favor of

- port-comb-srt
- port-present-srt
- port-cam-srt

This drops support for paths and params in srt URIs, whis is up for debate but in my oppinion not needed for our usecase.